requirements.txt
streamlit==1.7.0
h5py==3.1.0
numpy==1.19.5
Pillow==8.3.2
tensorflow-cpu==2.7.0
simplejson
